auto check for admin in _extras.ps1

+++ _extras.ps1
@@ -1,3 +1,18 @@
+$currentPrincipal = New-Object Security.Principal.WindowsPrincipal([Security.Principal.WindowsIdentity]::GetCurrent())
+$isAdmin = $currentPrincipal.IsInRole([Security.Principal.WindowsBuiltInRole]::Administrator)
+if (!$isAdmin) {
+	echo "This script must be run as administrator!"
+	
+	if (gcm "sudo" -ErrorAction SilentlyContinue) {
+		$confirmation = Read-Host "Do you want to run the script with elevated privileges? [Y/n] "
+		if ($confirmation -ne 'n') {
+		  sudo $MyInvocation.MyCommand.Source
+			exit 0
+		}
+	}
+	exit 1
+}
+
 # extra commandline utils
 scoop install imagemagick python python2 bat caddy cmder dos2unix ffmpeg jq php sed
